Implement auth using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -68,7 +68,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view("questions.edit", compact('question'));
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access Denied");
+        }
+            return view("questions.edit", compact('question'));
     }
 
     /**
@@ -80,6 +83,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access Denied");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', 'Your Question has been updated.');
@@ -93,6 +100,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "Access Denied");
+        }
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your Question has been deleted.');
